Move tests for deprecated extensionhelper to component
Followup (forgot to do it) after https://github.com/open-telemetry/opentelemetry-collector/pull/4805

Updates https://github.com/open-telemetry/opentelemetry-collector/issues/4681
